misc: Organize ViewModels
