+++ src/main/java/com/atlauncher/viewmodel/base/INewsViewModel.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package com.atlauncher.gui.tabs.news;
+package com.atlauncher.viewmodel.base;
 
 import java.util.function.Consumer;
